Update Data
Solid implementation but there was some random issue it looked like with syntax...I went through and redid the breaks and spacing to make sure it was all uniform and proper tabbing. Also added print statements and redid the dictionary accumulation process as it seemed a bit messy and off. This should work now!